rename assets folder images
