+++ hello-vite/src/images/chicken.jpg
